Read 2-channel PWM input, make it available on SPI
I can't believe I got this to work with such small jitter.  It bounces
around by about 10-15 counts but is usually pretty dead-on.  I might do
a histogram and add an input filter.

Something is a little funky electrically: it only works if pullups are
enabled, which leads me to believe either the radio is open collector or
its output impedance is pretty high.  Nevertheless, with my 10k/10k
resistor divider, it works.  I need a scope to see what's actually going
on, but I don't think I can possibly be exceeding 3.3V on the input
pins at least.